# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateBookingRequest {
  /** @minLength 3 */
  name: string;
  /** @minLength 10 */
  phone: string;
  service: string;
  message?: string;
}

export type BookingResponseStatus = typeof BookingResponseStatus[keyof typeof BookingResponseStatus];


export const BookingResponseStatus = {
  new: 'new',
  contacted: 'contacted',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface BookingResponse {
  id: number;
  name: string;
  phone: string;
  service: string;
  message?: string | null;
  status: BookingResponseStatus;
  createdAt: string;
}

export interface BookingsListResponse {
  bookings: BookingResponse[];
  total: number;
}

export type UpdateBookingStatusRequestStatus = typeof UpdateBookingStatusRequestStatus[keyof typeof UpdateBookingStatusRequestStatus];


export const UpdateBookingStatusRequestStatus = {
  new: 'new',
  contacted: 'contacted',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface UpdateBookingStatusRequest {
  status: UpdateBookingStatusRequestStatus;
}

export interface ErrorResponse {
  error: string;
}