# lib/api-zod/src/generated/types/bookingResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingResponseStatus } from './bookingResponseStatus';

export interface BookingResponse {
  id: number;
  name: string;
  phone: string;
  service: string;
  message?: string | null;
  status: BookingResponseStatus;
  createdAt: Date;
}