# lib/api-zod/src/generated/types/bookingsListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingResponse } from './bookingResponse';

export interface BookingsListResponse {
  bookings: BookingResponse[];
  total: number;
}