# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  /** @minLength 3 */
  name: string;
  /** @minLength 10 */
  phone: string;
  service: string;
  message?: string;
}